requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.11.2
certifi==2023.7.22
charset-normalizer==3.3.1
defusedxml==0.7.1
Django==4.2.6
django-appconf==1.0.5
django-axes==6.1.1
django-crum==0.7.9
django-dbbackup==4.0.2
django-filter==23.3
django-imagekit==5.0.0
django-ipware==5.0.2
django-modelcluster==6.1
django-permissionedforms==0.1
django-taggit==4.0.0
django-treebeard==4.7
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==21.2.0
html5lib==1.1
idna==3.4
l18n==2021.3
openpyxl==3.1.2
packaging==23.2
pilkit==3.0
Pillow==10.1.0
pillow-heif==0.13.1
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
typing_extensions==4.8.0
urllib3==2.0.7
wagtail==5.1.3
wagtail-modeladmin==1.0.0
wagtailgeowidget==8.0.0
webencodings==0.5.1
Willow==1.6.2
